requirements.txt
numpy
opencv-python
Pillow
onnx
onnxruntime
tqdm
scikit-image
torch